+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootBlaze.java
@@ -1,27 +1,27 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.WitherSkeletonEntity;
+import net.minecraft.entity.monster.BlazeEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootWitherskeleton 
+public class CustomLootBlaze 
 {
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof WitherSkeletonEntity) {
-            if(random.nextInt(3) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(4)));
+        if(event.getEntityLiving() instanceof BlazeEntity) {
             if(random.nextInt(6) == 0)
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
+            if(random.nextInt(12) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
             
-            if(random.nextInt(40) == 0)
+            if(random.nextInt(60) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_power_shard,1));
             
 	        }
