+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootCreeper.java
@@ -1,23 +1,23 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.EndermiteEntity;
+import net.minecraft.entity.monster.CreeperEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootEndermite 
+public class CustomLootCreeper 
 {
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof EndermiteEntity) {
-            if(random.nextInt(2) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee,2));
+        if(event.getEntityLiving() instanceof CreeperEntity) {
+            if(random.nextInt(4) == 0)
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
             if(random.nextInt(7) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
