+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootEndermite.java
@@ -1,24 +1,24 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.ZombieVillagerEntity;
+import net.minecraft.entity.monster.EndermiteEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootZombievillager 
+public class CustomLootEndermite 
 {
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof ZombieVillagerEntity) {
+        if(event.getEntityLiving() instanceof EndermiteEntity) {
+            if(random.nextInt(2) == 0)
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee,2));
             if(random.nextInt(7) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
-            if(random.nextInt(12) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
             
 	        }
