+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootGhast.java
@@ -1,27 +1,27 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.ZombifiedPiglinEntity;
+import net.minecraft.entity.monster.GhastEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootPigZombie 
+public class CustomLootGhast 
 {
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof ZombifiedPiglinEntity) {
-            if(random.nextInt(6) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
-            if(random.nextInt(13) == 0)
+        if(event.getEntityLiving() instanceof GhastEntity) {
+            if(random.nextInt(2) == 0)
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(4)));
+            if(random.nextInt(4) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
             
-            if(random.nextInt(70) == 0)
+            if(random.nextInt(35) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_power_shard,1));
             
 	        }
