+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootHusk.java
@@ -1,28 +1,25 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.VindicatorEntity;
+import net.minecraft.entity.monster.HuskEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootVindicator {
-	
+public class CustomLootHusk 
+{
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof VindicatorEntity) {
-            if(random.nextInt(2) == 0)
+        if(event.getEntityLiving() instanceof HuskEntity) {
+            if(random.nextInt(6) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
-            if(random.nextInt(5) == 0)
+            if(random.nextInt(12) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
-            
-            if(random.nextInt(25) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_courage_shard,1));
             
 	        }
 	    }
