+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootPigZombie.java
@@ -1,28 +1,28 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.DrownedEntity;
+import net.minecraft.entity.monster.ZombifiedPiglinEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootDrowned 
+public class CustomLootPigZombie 
 {
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof DrownedEntity) {
+        if(event.getEntityLiving() instanceof ZombifiedPiglinEntity) {
             if(random.nextInt(6) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
-            if(random.nextInt(14) == 0)
+            if(random.nextInt(13) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
             
             if(random.nextInt(70) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_courage_shard,1));
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_power_shard,1));
             
 	        }
 	    }
