+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootShulker.java
@@ -1,27 +1,27 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.PhantomEntity;
+import net.minecraft.entity.monster.ShulkerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootPhantom 
+public class CustomLootShulker 
 {
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof PhantomEntity) {
-            if(random.nextInt(5) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(4)));
-            if(random.nextInt(10) == 0)
+        if(event.getEntityLiving() instanceof ShulkerEntity) {
+            if(random.nextInt(3) == 0)
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
+            if(random.nextInt(7) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
             
-            if(random.nextInt(45) == 0)
+            if(random.nextInt(25) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_wisdom_shard,1));
             
 	        }
