+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootSilverfish.java
@@ -1,24 +1,24 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.CreeperEntity;
+import net.minecraft.entity.monster.SilverfishEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootCreeper 
+public class CustomLootSilverfish 
 {
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof CreeperEntity) {
+        if(event.getEntityLiving() instanceof SilverfishEntity) {
             if(random.nextInt(4) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
-            if(random.nextInt(7) == 0)
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee,1));
+            if(random.nextInt(15) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
             
 	        }
