+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootSkeleton.java
@@ -1,28 +1,28 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.BlazeEntity;
+import net.minecraft.entity.monster.SkeletonEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootBlaze 
+public class CustomLootSkeleton 
 {
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof BlazeEntity) {
-            if(random.nextInt(6) == 0)
+        if(event.getEntityLiving() instanceof SkeletonEntity) {
+            if(random.nextInt(5) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
-            if(random.nextInt(12) == 0)
+            if(random.nextInt(16) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
             
-            if(random.nextInt(60) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_power_shard,1));
+            if(random.nextInt(75) == 0)
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_wisdom_shard,1));
             
 	        }
 	    }
