+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootSlime.java
@@ -1,32 +1,32 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
 import net.minecraft.entity.monster.MagmaCubeEntity;
+import net.minecraft.entity.monster.SlimeEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootMagmacube 
+public class CustomLootSlime 
 {
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof MagmaCubeEntity) {
-            if(random.nextInt(8) == 0)
+        if(event.getEntityLiving() instanceof SlimeEntity && !(event.getEntityLiving() instanceof MagmaCubeEntity)) {
+            if(random.nextInt(9) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
-            if(random.nextInt(10) == 0)
+            if(random.nextInt(11) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
+            if(random.nextInt(14) == 0)
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.green_jelly,1));
             
-            if(random.nextInt(90) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_power_shard,1));
-            if(random.nextInt(20) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.red_jelly,1));
-            if(random.nextInt(60) == 0)
+            if(random.nextInt(65) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_jelly,1));
+            
 	        }
 	    }
 	}
