+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootVindicator.java
@@ -1,28 +1,28 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.GhastEntity;
+import net.minecraft.entity.monster.VindicatorEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootGhast 
-{
+public class CustomLootVindicator {
+	
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof GhastEntity) {
+        if(event.getEntityLiving() instanceof VindicatorEntity) {
             if(random.nextInt(2) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(4)));
-            if(random.nextInt(4) == 0)
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
+            if(random.nextInt(5) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
             
-            if(random.nextInt(35) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_power_shard,1));
+            if(random.nextInt(25) == 0)
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_courage_shard,1));
             
 	        }
 	    }
