+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootWitch.java
@@ -1,25 +1,28 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.EndermanEntity;
+import net.minecraft.entity.monster.WitchEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootEnderman 
+public class CustomLootWitch 
 {
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof EndermanEntity) {
-            if(random.nextInt(3) == 0)
+        if(event.getEntityLiving() instanceof WitchEntity) {
+            if(random.nextInt(2) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(4)));
-            if(random.nextInt(8) == 0)
+            if(random.nextInt(4) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
+            
+            if(random.nextInt(20) == 0)
+                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_wisdom_shard,1));
             
 	        }
 	    }
