+++ src/main/java/superworldsun/superslegend/customlootmobs/CustomLootZombievillager.java
@@ -1,28 +1,25 @@
-package superworldsun.superslegend.CustomLootMobs;
+package superworldsun.superslegend.customlootmobs;
 
 import java.util.Random;
 
-import net.minecraft.entity.monster.GuardianEntity;
+import net.minecraft.entity.monster.ZombieVillagerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingDropsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import superworldsun.superslegend.lists.ItemList;
 
-public class CustomLootGuardian 
+public class CustomLootZombievillager 
 {
 	@SubscribeEvent
     public void CustomLootDrops(LivingDropsEvent event) {
 
         Random random = new Random();
 
-        if(event.getEntityLiving() instanceof GuardianEntity) {
-            if(random.nextInt(4) == 0)
+        if(event.getEntityLiving() instanceof ZombieVillagerEntity) {
+            if(random.nextInt(7) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.rupee, random.nextInt(3)));
-            if(random.nextInt(6) == 0)
+            if(random.nextInt(12) == 0)
                 event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.blue_rupee,1));
-            
-            if(random.nextInt(55) == 0)
-                event.getEntityLiving().spawnAtLocation(new ItemStack(ItemList.triforce_courage_shard,1));
             
 	        }
 	    }
